requirements.txt
coreapi==2.3.3
Django==2.1.7
djangorestframework==3.9.1